requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Limiter==3.8.0
Flask-CORS==5.0.1
Flask-Mail==0.10.0
bcrypt==4.2.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
